requirements.txt
celery==5.4.0 # Period and async tasks
crispy-bootstrap5==2024.10 # Template pack for Crispy forms
boto3==1.20.11 # Used for Results JSON Uploads to S3
botocore==1.23.11 # Pre-req for Above
certifi==2024.7.4 # Used by Requests, Mozilla's CA Bundle
Django~=4.2.0 # Django
django-allauth==65.3.1 # Auth and 3rd party login support
django-celery-beat==2.5.0 # Allows storing periodic task schedule in the database, passed as an arg to Celery worker
django-crispy-forms==2.3.0 # Easier Forms
django-dbbackup==4.0.1 # Util to enable easy backups from the CLI - Used by Celery
django-debug-toolbar==4.4.6 # Debug Toolbar
django-environ==0.12.0 # Easy config of things like DB from Env Vars
django-extensions==3.2.3 # Extra admin CLI tools
django-filter==2.4.0 # Filter QuerySets with URL Query Params
django-heroku==0.3.1 # Archived - Consider removing
django-inlinecss-redux==0.4.0 # Fork of original Inline CSS - Used for Email Generation
django-redirect-to-non-www==0.1.1 # Might Not Be Needed?
django-redis==4.12.1 # Enables Redis as a cache backend for Django
django-ses==3.5.0 # Django Email Backend for Amazon SES
django-storages==1.14.4 # Storage backend, support for S3 through Boto
django-timezone-field==5.0.0 # Timezone Model Field
djangorestframework==3.15.2 # DRF
djangorestframework-csv==3.0.2 #CSV Output for DRF
gunicorn==22.0.0 # Webserver
Markdown==3.3.4 # Markdown Support - Used by Blog
Pillow==10.3.0 # Needed for ImageField Support on Models
psycopg[binary]==3.2.3 # Python PostGres support, used by Django ORM
pytz==2019.3 # Timezones <- Depracated, Switch to zoneinfo for Django 5 onwards
redis==3.5.3 # Caching and Celery Brokerage
requests==2.32.2 # HTTP Client Used by Django
sentry-sdk==2.8.0 # Sentry Logging
urllib3==1.26.19 # Dependency for Requests - might not be needed?
django-viewflow==2.2.9 # Replaces Material Library, which used "smart_text" (unsupported as of Django 4.0)
whitenoise==6.8.2 # Enable Static Files from Webserver